Add `Sai::Decorator::NamedStyles`
